backend_service/requirements.txt
Flask>=2.3.0
Flask-CORS>=4.0.0
requests>=2.30.0
groq
langchain
langchain-core
langchain-community
langchain-huggingface
huggingface-hub>=1.0.0
langchain-text-splitters
sentence-transformers
python-dotenv
chromadb
langchain-groq
langchain-chroma
docx2txt